sua minvo

+++ src/main/java/org/example/demo/Entity.java
@@ -1,6 +1,8 @@
 package org.example.demo;
 
 public class Entity{
-    public int x,y;
+    public double x,y;
     public int speed;
+
+
 }
